requirements.txt
langchain
langchain-experimental
langchain_experimental
openai
python-dotenv
pymysql
langsmith
